pyproject.toml
[project]
name = "smart-web-attack-replay-generator"
version = "1.0.0"
description = "Smart Web Application Attack Replay Generator - Parse logs, detect attacks, generate replay scripts"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.51.0",
    "pandas>=2.0.0",
    "requests>=2.31.0",
    "psycopg2-binary>=2.9.11",
    "plotly>=6.5.0",
]
